policy: Policy Map Overflow Lockdown
Assuming `enable-lockdown-endpoint-on-policy-overflow`)
is set to "true" (by default it is "false"),when a
policy map needs more entries than is allocated for its
bpf map, the endpoint is locked down. The policy cache is
reset and the policy map is replaced with a set of default
deny entries. The endpoint is put into full default-deny
for every protocol (except for ICMP).

If the lockdown calls fail, then the endpoint will be
regenerated, and another lockdown will be attempted (until
the underlying map pressure is relieved).

Incremental deletion is now done first in cases where
incremental addition before deletion would overflow the
policy map.

Signed-off-by: Nate Sweet <nathanjsweet@pm.me>