pkg/tuple: make TupleKey4,6 both an alias of generic tupleKey type.
These types are identical, except for the ipv{4,6} type used.
This makes TupleKey{4,6} types both be type aliases of a generic
tupleKey[T] type.

As well, this adds getter functions to remaining fields in both ipv{4,6}
implementations to allow for a single interface to describe the actions
of accessing tuple fields (with all addr fields are converted to netip.Addr).

In subsequent commits, this will be used to DRY up code (i.e. ctmap) that
depends on these types such that we don't need to maintain two identical
code paths for each IP family, as is the case with ctmap GC.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>